+++ Contents/mods/pz-zombie-factions/42/media/lua/client/ZombieFactions/HordeSpawnerExtension_Client.lua
@@ -10,7 +10,7 @@ local REL = ZombieFactions.Relationship
 local originalCreateChildren = ISSpawnHordeUI.createChildren
 local originalOnSpawn = ISSpawnHordeUI.onSpawn
 
-print("[ZombieFactions] Client Horde Spawner extension loaded v0.0.29")
+print("[ZombieFactions] Client Horde Spawner extension loaded v0.0.30")
 
 local function addRelationshipOptions(combo)
     combo:addOptionWithData("FRIENDLY", REL.FRIENDLY)